Separating the application and the server
